[#12486] refactor(core): align user/group cache version check with project conventions
Move the helper under storage.relational.utils, keep private methods at the
end of RelationalEntityStore, order builders with fields, and drop core
Javadoc references to the JCasbin authorizer.
